+++ src/why/index.tsx
@@ -1,1 +1,0 @@
-export { Why, WhyProps } from './Why'